global state for degree and drop down
