+++ src/components/DropdownDegrees.tsx
@@ -1,5 +1,4 @@
-import React, { useState } from "react";
-import { AllDegrees } from "../interfaces/AllCourses-AllDegrees";
+import React from "react";
 import { Degree } from "../interfaces/course-Degree-Semester";
 import "../App.css";
 import { Form } from "react-bootstrap";
@@ -8,18 +7,26 @@ type ChangeEvent = React.ChangeEvent<
     HTMLTextAreaElement | HTMLInputElement | HTMLSelectElement
 >;
 
-export function DropdownDegrees(): JSX.Element {
-    const degrees = [...AllDegrees];
-    const [degree, setDegree] = useState<Degree>(degrees[0]);
-    function updateDegree(event: ChangeEvent) {
-        const newDegreeName = event.target.value;
-        const newDegree = degrees.find(
-            (degree: Degree): boolean => degree.name === newDegreeName
-        );
-        if (newDegree !== undefined) {
-            setDegree(newDegree);
-        }
-    }
+export function DropdownDegrees({
+    degrees,
+    degree,
+    updateDegree
+}: {
+    degrees: Degree[];
+    degree: Degree;
+    updateDegree: (event: ChangeEvent) => void;
+}): JSX.Element {
+    // const degrees = [...AllDegrees];
+    // const [degree, setDegree] = useState<Degree>(degrees[0]);
+    // function updateDegree(event: ChangeEvent) {
+    //     const newDegreeName = event.target.value;
+    //     const newDegree = degrees.find(
+    //         (degree: Degree): boolean => degree.name === newDegreeName
+    //     );
+    //     if (newDegree !== undefined) {
+    //         setDegree(newDegree);
+    //     }
+    // }
     return (
         <div>
             <Form.Group controlId="Degrees">
